Remove RequestSender and use HttpProtocol.proxyRequest
